563 LocalMove, moved FlipMove and StretchMove to new files
