+++ Assets/SEE/Layout/NodeLayouts/IncrementalTreeMap/LocalMove.cs
@@ -1,6 +1,4 @@
 using System.Collections.Generic;
-using System.Diagnostics;
-using System;
 
 namespace SEE.Layout.NodeLayouts.IncrementalTreeMap
 {
@@ -32,388 +30,5 @@ namespace SEE.Layout.NodeLayouts.IncrementalTreeMap
         /// of the node layout of of <see cref="Node1"/> and <see cref="Node2"/>.</param>
         /// <returns>a new local move</returns>
         public abstract LocalMove Clone(IDictionary<string, Node> cloneMap);
-    }
-
-    /// <summary>
-    /// A flip move is a kind of <see cref="LocalMove"/>.
-    /// It rotates two nodes 90 degrees.
-    /// </summary>
-    [DebuggerDisplay("{DebuggerDisplay,nq}")]
-    internal class FlipMove : LocalMove
-    {
-        /// <summary>
-        /// if the rotation is clockwise
-        /// </summary>
-        private readonly bool clockwise;
-
-        /// <summary>
-        /// Constructor
-        /// </summary>
-        /// <param name="node1">one affected node</param>
-        /// <param name="node2">other affected node</param>
-        /// <param name="clockwise">if the rotation is clockwise </param>
-        public FlipMove(Node node1, Node node2, bool clockwise)
-        {
-            this.Node1 = node1;
-            this.Node2 = node2;
-            this.clockwise = clockwise;
-        }
-
-        override public void Apply()
-        {
-            IDictionary<Direction, Segment> segmentsNode1 = Node1.SegmentsDictionary();
-            IDictionary<Direction, Segment> segmentsNode2 = Node2.SegmentsDictionary();
-
-            if (segmentsNode1[Direction.Right] == segmentsNode2[Direction.Left])
-            {
-                // [Node1][Node2]
-                FlipOnVerticalSegment(leftNode: Node1, rightNode: Node2);
-            }
-            else if (segmentsNode1[Direction.Left] == segmentsNode2[Direction.Right])
-            {
-                // [Node2][Node1]
-                FlipOnVerticalSegment(leftNode: Node2, rightNode: Node1);
-            }
-            else if (segmentsNode1[Direction.Upper] == segmentsNode2[Direction.Lower])
-            {
-                // [Node2]
-                // [Node1]
-                FlipOnHorizontalSegment(lowerNode: Node1, upperNode: Node2);
-            }
-            else if (segmentsNode1[Direction.Lower] == segmentsNode2[Direction.Upper])
-            {
-                // [Node1]
-                // [Node2]
-                FlipOnHorizontalSegment(lowerNode: Node2, upperNode: Node1);
-            }
-            else
-            {
-                throw new ArgumentException("Can't apply flip move");
-            }
-        }
-
-        override public LocalMove Clone(IDictionary<string, Node> cloneMap)
-        {
-            return new FlipMove(cloneMap[Node1.ID], cloneMap[Node2.ID], clockwise);
-        }
-
-        /// <summary>
-        /// Applies the local move for the case that the nodes are separated vertically
-        /// </summary>
-        /// <param name="leftNode">the node on the <see cref="Direction.Left"/> side</param>
-        /// <param name="rightNode">the node on the <see cref="Direction.Right"/> side</param>
-        private void FlipOnVerticalSegment(Node leftNode, Node rightNode)
-        {
-            // clockwise            anticlockwise
-            // [l][r] -> [lll]      [l][r] -> [rrr]
-            // [l][r]    [rrr]      [l][r]    [lll]
-
-            // adjust rectangles
-            double width = leftNode.Rectangle.Width + rightNode.Rectangle.Width;
-            double ratio = leftNode.Rectangle.Area() / (leftNode.Rectangle.Area() + rightNode.Rectangle.Area());
-
-            leftNode.Rectangle.Width = width;
-            rightNode.Rectangle.Width = width;
-            rightNode.Rectangle.X = leftNode.Rectangle.X;
-
-            leftNode.Rectangle.Depth *= ratio;
-            rightNode.Rectangle.Depth *= (1 - ratio);
-            if (clockwise)
-            {
-                leftNode.Rectangle.Z = rightNode.Rectangle.Z + rightNode.Rectangle.Depth;
-            }
-            else
-            {
-                rightNode.Rectangle.Z = leftNode.Rectangle.Z + leftNode.Rectangle.Depth;
-            }
-
-            // switch segments
-            Segment newRightSegmentForLeftNode = rightNode.SegmentsDictionary()[Direction.Right];
-            Segment newLeftSegmentForRightNode = leftNode.SegmentsDictionary()[Direction.Left];
-            Segment middle = leftNode.SegmentsDictionary()[Direction.Right];
-            middle.IsVertical = false;
-
-            leftNode.RegisterSegment(newRightSegmentForLeftNode, Direction.Right);
-            rightNode.RegisterSegment(newLeftSegmentForRightNode, Direction.Left);
-
-            if (clockwise)
-            {
-                leftNode.RegisterSegment(middle, Direction.Lower);
-                rightNode.RegisterSegment(middle, Direction.Upper);
-            }
-            else
-            {
-                leftNode.RegisterSegment(middle, Direction.Upper);
-                rightNode.RegisterSegment(middle, Direction.Lower);
-            }
-        }
-
-        /// <summary>
-        /// Applies the local move for the case that the nodes are separated horizontally.
-        /// </summary>
-        /// <param name="lowerNode">the node on the <see cref="Direction.Lower"/> side</param>
-        /// <param name="upperNode">the node on the <see cref="Direction.Upper"/> side</param>
-        private void FlipOnHorizontalSegment(Node lowerNode, Node upperNode)
-        {
-            // clockwise                anticlockwise
-            // [uuu]  ->  [l][u]        [uuu]  ->  [u][l]
-            // [lll]      [l][u]        [lll]      [u][l]
-
-            // adjust rectangles
-            double depth = lowerNode.Rectangle.Depth + upperNode.Rectangle.Depth;
-            double ratio = lowerNode.Rectangle.Area() / (lowerNode.Rectangle.Area() + upperNode.Rectangle.Area());
-
-            lowerNode.Rectangle.Depth = depth;
-            upperNode.Rectangle.Depth = depth;
-            upperNode.Rectangle.Z = lowerNode.Rectangle.Z;
-
-            lowerNode.Rectangle.Width *= ratio;
-            upperNode.Rectangle.Width *= (1 - ratio);
-            if (clockwise)
-            {
-                upperNode.Rectangle.X = lowerNode.Rectangle.X + lowerNode.Rectangle.Width;
-            }
-            else
-            {
-                lowerNode.Rectangle.X = upperNode.Rectangle.X + upperNode.Rectangle.Width;
-            }
-
-            // switch segments
-            Segment newUpperSegmentForLowerNode = upperNode.SegmentsDictionary()[Direction.Upper];
-            Segment newLowerSegmentForUpperNode = lowerNode.SegmentsDictionary()[Direction.Lower];
-            Segment middle = lowerNode.SegmentsDictionary()[Direction.Upper];
-            middle.IsVertical = true;
-
-            lowerNode.RegisterSegment(newUpperSegmentForLowerNode, Direction.Upper);
-            upperNode.RegisterSegment(newLowerSegmentForUpperNode, Direction.Lower);
-
-            if (clockwise)
-            {
-                lowerNode.RegisterSegment(middle, Direction.Right);
-                upperNode.RegisterSegment(middle, Direction.Left);
-            }
-            else
-            {
-                lowerNode.RegisterSegment(middle, Direction.Left);
-                upperNode.RegisterSegment(middle, Direction.Right);
-            }
-        }
-
-        /// <summary>
-        /// Method for better overview in debugger
-        /// </summary>
-        private string DebuggerDisplay => "flip " + Node1.ID + " " + Node2.ID + " {clockwise}";
-    }
-
-    /// <summary>
-    /// A stretch move is a kind of <see cref="LocalMove"/>.
-    /// It expands a node over another.
-    /// </summary>
-    [DebuggerDisplay("{DebuggerDisplay,nq}")]
-    internal class StretchMove : LocalMove
-    {
-        /// <summary>
-        /// Constructor
-        /// </summary>
-        /// <param name="node1">one affected node</param>
-        /// <param name="node2">the other affected node</param>
-        public StretchMove(Node node1, Node node2)
-        {
-            Node1 = node1;
-            Node2 = node2;
-        }
-
-        override public void Apply()
-        {
-            IDictionary<Direction, Segment> segmentsNode1 = Node1.SegmentsDictionary();
-            IDictionary<Direction, Segment> segmentsNode2 = Node2.SegmentsDictionary();
-            if (segmentsNode1[Direction.Right] == segmentsNode2[Direction.Left]
-                || segmentsNode1[Direction.Left] == segmentsNode2[Direction.Right])
-            {
-                (Node leftNode, Node rightNode) = (segmentsNode1[Direction.Right] == segmentsNode2[Direction.Left])
-                    ? (Node1, Node2)
-                    : (Node2, Node1);
-                if (leftNode.Rectangle.Depth >= rightNode.Rectangle.Depth)
-                {
-                    StretchRightOverVertical(leftNode: leftNode, rightNode: rightNode);
-                }
-                else
-                {
-                    StretchLeftOverVertical(leftNode: leftNode, rightNode: rightNode);
-                }
-            }
-            else if (segmentsNode1[Direction.Upper] == segmentsNode2[Direction.Lower]
-                     || segmentsNode1[Direction.Lower] == segmentsNode2[Direction.Upper])
-            {
-                (Node lowerNode, Node upperNode) = (segmentsNode1[Direction.Upper] == segmentsNode2[Direction.Lower])
-                    ? (Node1, Node2)
-                    : (Node2, Node1);
-                if (lowerNode.Rectangle.Width >= upperNode.Rectangle.Width)
-                {
-                    StretchUpperOverHorizontal(lowerNode: lowerNode, upperNode: upperNode);
-                }
-                else
-                {
-                    StretchLowerOverHorizontal(lowerNode: lowerNode, upperNode: upperNode);
-                }
-            }
-            else
-            {
-                throw new ArgumentException("Can't apply stretch move");
-            }
-        }
-
-        override public LocalMove Clone(IDictionary<string, Node> cloneMap)
-        {
-            return new StretchMove(cloneMap[Node1.ID], cloneMap[Node2.ID]);
-        }
-
-        /// <summary>
-        /// Apply for the case that:
-        /// - the nodes are seperated vertically
-        /// - the left node is the expanding node
-        /// </summary>
-        /// <param name="leftNode">node on the <see cref="Direction.Left"/> side</param>
-        /// <param name="rightNode">node on the <see cref="Direction.Right"/> side</param>
-        private void StretchLeftOverVertical(Node leftNode, Node rightNode)
-        {
-            // along lower           along upper
-            //    [r]         [r]    [l][r]  ->  [llll]
-            // [l][r]  ->  [llll]       [r]         [r]
-            IDictionary<Direction, Segment> segmentsLeftNode = leftNode.SegmentsDictionary();
-            IDictionary<Direction, Segment> segmentsRightNode = rightNode.SegmentsDictionary();
-            bool alongLowerSegment = segmentsLeftNode[Direction.Lower] == segmentsRightNode[Direction.Lower];
-            // adjust rectangles
-            leftNode.Rectangle.Width += rightNode.Rectangle.Width;
-            rightNode.Rectangle.Depth -= leftNode.Rectangle.Depth;
-            if (alongLowerSegment)
-            {
-                rightNode.Rectangle.Z = leftNode.Rectangle.Z + leftNode.Rectangle.Depth;
-            }
-
-            // switch segments
-            leftNode.RegisterSegment(segmentsRightNode[Direction.Right], Direction.Right);
-            if (alongLowerSegment)
-            {
-                rightNode.RegisterSegment(segmentsLeftNode[Direction.Upper], Direction.Lower);
-            }
-            else
-            {
-                rightNode.RegisterSegment(segmentsLeftNode[Direction.Lower], Direction.Upper);
-            }
-        }
-
-        /// <summary>
-        /// Apply for the case that:
-        /// - the nodes are seperated vertically
-        /// - the right node is the expanding node
-        /// </summary>
-        /// <param name="leftNode">node on the <see cref="Direction.Left"/> side</param>
-        /// <param name="rightNode">node on the <see cref="Direction.Right"/> side</param>
-        private void StretchRightOverVertical(Node leftNode, Node rightNode)
-        {
-            // along lower           along upper
-            // [l]         [l]       [l][r]  ->  [r][r]
-            // [l][r]  ->  [r][r]    [l]         [l]
-            IDictionary<Direction, Segment> segmentsLeftNode = leftNode.SegmentsDictionary();
-            IDictionary<Direction, Segment> segmentsRightNode = rightNode.SegmentsDictionary();
-            bool alongLowerSegment = segmentsLeftNode[Direction.Lower] == segmentsRightNode[Direction.Lower];
-            // adjust rectangles
-            rightNode.Rectangle.Width += leftNode.Rectangle.Width;
-            rightNode.Rectangle.X = leftNode.Rectangle.X;
-            leftNode.Rectangle.Depth -= rightNode.Rectangle.Depth;
-            if (alongLowerSegment)
-            {
-                leftNode.Rectangle.Z = rightNode.Rectangle.Z + rightNode.Rectangle.Depth;
-            }
-
-            // switch segments
-            rightNode.RegisterSegment(segmentsLeftNode[Direction.Left], Direction.Left);
-            if (alongLowerSegment)
-            {
-                leftNode.RegisterSegment(segmentsRightNode[Direction.Upper], Direction.Lower);
-            }
-            else
-            {
-                leftNode.RegisterSegment(segmentsRightNode[Direction.Lower], Direction.Upper);
-            }
-        }
-
-        /// <summary>
-        /// Apply for the case that:
-        /// - the nodes are seperated horizontally
-        /// - the lower node is the expanding node
-        /// </summary>
-        /// <param name="lowerNode">node on the <see cref="Direction.Lower"/> side</param>
-        /// <param name="upperNode">node on the <see cref="Direction.Upper"/> side</param>
-        private void StretchLowerOverHorizontal(Node lowerNode, Node upperNode)
-        {
-            // along left           along right
-            // [uuuu] ->  [l][u]    [uuuu]  ->  [u][l]
-            // [l]        [l]          [l]         [l]
-            IDictionary<Direction, Segment> segmentsLowerNode = lowerNode.SegmentsDictionary();
-            IDictionary<Direction, Segment> segmentsUpperNode = upperNode.SegmentsDictionary();
-            bool alongLeftSegment = segmentsLowerNode[Direction.Left] == segmentsUpperNode[Direction.Left];
-            // adjust rectangles
-            lowerNode.Rectangle.Depth += upperNode.Rectangle.Depth;
-            upperNode.Rectangle.Width -= lowerNode.Rectangle.Width;
-            if (alongLeftSegment)
-            {
-                upperNode.Rectangle.X = lowerNode.Rectangle.X + lowerNode.Rectangle.Width;
-            }
-
-            // switch segments
-            lowerNode.RegisterSegment(segmentsUpperNode[Direction.Upper], Direction.Upper);
-            if (alongLeftSegment)
-            {
-                upperNode.RegisterSegment(segmentsLowerNode[Direction.Right], Direction.Left);
-            }
-            else
-            {
-                upperNode.RegisterSegment(segmentsLowerNode[Direction.Left], Direction.Right);
-            }
-        }
-
-        /// <summary>
-        /// Apply for the case that:
-        /// - the nodes are seperated horizontally
-        /// - the upper node is the expanding node
-        /// </summary>
-        /// <param name="lowerNode">node on the <see cref="Direction.Lower"/> side</param>
-        /// <param name="upperNode">node on the <see cref="Direction.Upper"/> side</param>
-        private void StretchUpperOverHorizontal(Node lowerNode, Node upperNode)
-        {
-            // along left           along right
-            // [u]    ->  [u]          [u]  ->      [u]
-            // [llll]     [u][l]    [llll]       [l][u]
-            IDictionary<Direction, Segment> segmentsLowerNode = lowerNode.SegmentsDictionary();
-            IDictionary<Direction, Segment> segmentsUpperNode = upperNode.SegmentsDictionary();
-            bool alongLeftSegment = segmentsLowerNode[Direction.Left] == segmentsUpperNode[Direction.Left];
-            // adjust rectangles
-            upperNode.Rectangle.Depth += lowerNode.Rectangle.Depth;
-            upperNode.Rectangle.Z = lowerNode.Rectangle.Z;
-
-            lowerNode.Rectangle.Width -= upperNode.Rectangle.Width;
-            if (alongLeftSegment)
-            {
-                lowerNode.Rectangle.X = upperNode.Rectangle.X + upperNode.Rectangle.Width;
-            }
-
-            // switch segments
-            upperNode.RegisterSegment(segmentsLowerNode[Direction.Lower], Direction.Lower);
-            if (alongLeftSegment)
-            {
-                lowerNode.RegisterSegment(segmentsUpperNode[Direction.Right], Direction.Left);
-            }
-            else
-            {
-                lowerNode.RegisterSegment(segmentsUpperNode[Direction.Left], Direction.Right);
-            }
-        }
-
-        /// <summary>
-        /// Method for better overview in debugger
-        /// </summary>
-        private string DebuggerDisplay => "stretch " + Node1.ID + " " + Node2.ID;
     }
 }
